Something like a launcher
Game can be started from Startup.java, which is a very basic launcher
giving the user the option to choose between the "main game" and our
physics test. Meant to evolve into something like a real launcher. I'd
love to have one that looks like the launcher of TES V: Skyrim.
Furthermore, there is a new restart button in the in-game menu.